Pull up artifact/metadata max ages to proxy repo
And remove all the duplicated code. No upgrade needed,
as the data remained where it was, merely the code handling
it was pulled up a level.
