+++ components/nexus-core/src/main/java/org/sonatype/nexus/proxy/maven/AbstractMavenRepository.java
@@ -276,26 +276,6 @@ public abstract class AbstractMavenRepository
   }
 
   @Override
-  public int getArtifactMaxAge() {
-    return getExternalConfiguration(false).getArtifactMaxAge();
-  }
-
-  @Override
-  public void setArtifactMaxAge(int maxAge) {
-    getExternalConfiguration(true).setArtifactMaxAge(maxAge);
-  }
-
-  @Override
-  public int getMetadataMaxAge() {
-    return getExternalConfiguration(false).getMetadataMaxAge();
-  }
-
-  @Override
-  public void setMetadataMaxAge(int metadataMaxAge) {
-    getExternalConfiguration(true).setMetadataMaxAge(metadataMaxAge);
-  }
-
-  @Override
   public boolean isMavenArtifact(StorageItem item) {
     return isMavenArtifactPath(item.getPath());
   }
