+++ components/nexus-core/src/main/java/org/sonatype/nexus/proxy/repository/AbstractProxyRepositoryConfiguration.java
@@ -23,6 +23,10 @@ public abstract class AbstractProxyRepositoryConfiguration
 
   private static final String ITEM_MAX_AGE = "itemMaxAge";
 
+  public static final String ARTIFACT_MAX_AGE = "artifactMaxAge";
+
+  public static final String METADATA_MAX_AGE = "metadataMaxAge";
+
   private static final String ITEM_AGING_ACTIVE = "itemAgingActive";
 
   private static final String AUTO_BLOCK_ACTIVE = "autoBlockActive";
@@ -65,6 +69,22 @@ public abstract class AbstractProxyRepositoryConfiguration
 
   public void setItemMaxAge(int age) {
     setNodeValue(getRootNode(), ITEM_MAX_AGE, String.valueOf(age));
+  }
+
+  public int getArtifactMaxAge() {
+    return Integer.parseInt(getNodeValue(getRootNode(), ARTIFACT_MAX_AGE, "1440"));
+  }
+
+  public void setArtifactMaxAge(int age) {
+    setNodeValue(getRootNode(), ARTIFACT_MAX_AGE, String.valueOf(age));
+  }
+
+  public int getMetadataMaxAge() {
+    return Integer.parseInt(getNodeValue(getRootNode(), METADATA_MAX_AGE, "1440"));
+  }
+
+  public void setMetadataMaxAge(int age) {
+    setNodeValue(getRootNode(), METADATA_MAX_AGE, String.valueOf(age));
   }
 
   public boolean isItemAgingActive() {
