+++ plugins/p2/nexus-p2-repository-plugin/src/main/java/org/sonatype/nexus/plugins/p2/repository/P2ProxyRepository.java
@@ -32,14 +32,6 @@ public interface P2ProxyRepository
   Map<String, ArtifactMapping> getArtifactMappings()
       throws IllegalOperationException, StorageException;
 
-  int getArtifactMaxAge();
-
-  void setArtifactMaxAge(final int maxAge);
-
-  int getMetadataMaxAge();
-
-  void setMetadataMaxAge(final int metadataMaxAge);
-
   ChecksumPolicy getChecksumPolicy();
 
   void setChecksumPolicy(final ChecksumPolicy checksumPolicy);
