+++ plugins/p2/nexus-p2-repository-plugin/src/main/java/org/sonatype/nexus/plugins/p2/repository/UpdateSiteProxyRepository.java
@@ -18,12 +18,4 @@ public interface UpdateSiteProxyRepository
     extends ProxyRepository
 {
   void mirror(final boolean force);
-
-  int getArtifactMaxAge();
-
-  void setArtifactMaxAge(final int maxAge);
-
-  int getMetadataMaxAge();
-
-  void setMetadataMaxAge(final int metadataMaxAge);
 }
