+++ plugins/p2/nexus-p2-repository-plugin/src/main/java/org/sonatype/nexus/plugins/p2/repository/updatesite/UpdateSiteRepositoryConfiguration.java
@@ -19,27 +19,7 @@ import org.codehaus.plexus.util.xml.Xpp3Dom;
 public class UpdateSiteRepositoryConfiguration
     extends AbstractProxyRepositoryConfiguration
 {
-  public static final String ARTIFACT_MAX_AGE = "artifactMaxAge";
-
-  public static final String METADATA_MAX_AGE = "metadataMaxAge";
-
   public UpdateSiteRepositoryConfiguration(final Xpp3Dom configuration) {
     super(configuration);
-  }
-
-  public int getArtifactMaxAge() {
-    return Integer.parseInt(getNodeValue(getRootNode(), ARTIFACT_MAX_AGE, "1440"));
-  }
-
-  public void setArtifactMaxAge(final int age) {
-    setNodeValue(getRootNode(), ARTIFACT_MAX_AGE, String.valueOf(age));
-  }
-
-  public int getMetadataMaxAge() {
-    return Integer.parseInt(getNodeValue(getRootNode(), METADATA_MAX_AGE, "1440"));
-  }
-
-  public void setMetadataMaxAge(final int age) {
-    setNodeValue(getRootNode(), METADATA_MAX_AGE, String.valueOf(age));
   }
 }
