+++ plugins/rubygem/nexus-ruby-plugin/src/main/java/org/sonatype/nexus/plugins/ruby/proxy/DefaultProxyRubyRepositoryConfiguration.java
@@ -24,10 +24,6 @@ import org.codehaus.plexus.util.xml.Xpp3Dom;
 public class DefaultProxyRubyRepositoryConfiguration
     extends AbstractProxyRepositoryConfiguration
 {
-  public static final String ARTIFACT_MAX_AGE = "artifactMaxAge";
-
-  public static final String METADATA_MAX_AGE = "metadataMaxAge";
-
   public DefaultProxyRubyRepositoryConfiguration(Xpp3Dom configuration) {
     super(configuration);
   }
@@ -36,15 +32,7 @@ public class DefaultProxyRubyRepositoryConfiguration
     return Integer.parseInt(getNodeValue(getRootNode(), ARTIFACT_MAX_AGE, "-1"));
   }
 
-  public void setArtifactMaxAge(int age) {
-    setNodeValue(getRootNode(), ARTIFACT_MAX_AGE, String.valueOf(age));
-  }
-
   public int getMetadataMaxAge() {
     return Integer.parseInt(getNodeValue(getRootNode(), METADATA_MAX_AGE, "30"));
-  }
-
-  public void setMetadataMaxAge(int age) {
-    setNodeValue(getRootNode(), METADATA_MAX_AGE, String.valueOf(age));
   }
 }
